+++ tidy3d/components/simulation.py
@@ -527,11 +527,11 @@ class Simulation(Box):
             if isinstance(structure.medium, Medium2D):
                 continue
             for geom in flatten_groups(structure.geometry):
-                zero_axes = [ind for ind, ele in enumerate(geom.bounding_box.size) if ele == 0.0]
-                if len(zero_axes) > 0:
+                zero_dims = geom.zero_dims
+                if len(zero_dims) > 0:
                     log.warning(
                         f"Structure at 'structures[{i}]' has geometry with zero size along "
-                        f"dimensions {zero_axes}, and with a medium that is not a 'Medium2D'. "
+                        f"dimensions {zero_dims}, and with a medium that is not a 'Medium2D'. "
                         "This is probably not correct, since the resulting simulation will "
                         "depend on the details of the numerical grid. Consider either "
                         "giving the geometry a nonzero thickness or using a 'Medium2D'."