+++ tidy3d/components/geometry/utils.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 from typing import Union, Tuple
 
-from ..types import Axis, PlanePosition, Shapely
+from ..types import Axis, PlanePosition, Shapely, ArrayFloat2D
 from ...exceptions import Tidy3dError
 
 from . import base
@@ -155,5 +155,29 @@ def from_shapely(
                 for geo in shape.geoms
             ]
         )
+
+    raise Tidy3dError(f"Shape {shape} cannot be converted to Geometry.")
+
+
+def vertices_from_shapely(shape: Shapely) -> ArrayFloat2D:
+    """Iterate over the polygons of a shapely geometry returning the vertices.
+
+    Parameters
+    ----------
+    shape : shapely.geometry.base.BaseGeometry
+        Shapely primitive to have its vertices extracted. It must be a linear ring, a polygon or a
+        collection of any of those.
+
+    Returns
+    -------
+    List[Tuple[ArrayFloat2D]]
+        List of tuples `(exterior, *interiors)`.
+    """
+    if shape.geom_type == "LinearRing":
+        return [(shape.coords[:-1],)]
+    if shape.geom_type == "Polygon":
+        return [(shape.exterior.coords[:-1],) + tuple(hole.coords[:-1] for hole in shape.interiors)]
+    if shape.geom_type in {"MultiPolygon", "GeometryCollection"}:
+        return sum(vertices_from_shapely(geo) for geo in shape.geoms)
 
     raise Tidy3dError(f"Shape {shape} cannot be converted to Geometry.")