+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Properly handle `.freqs` in `output_monitors` of adjoint plugin.
+- Simulation updater for `.hdf5` files with custom data.
 
 ## [2.4.2] - 2023-9-28